# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.roleCreated:
    properties:
      role_name:
        type: string
      role_ref:
        type: string
    type: object
...
  models.GroupRolePermission:
    properties:
      model:
        type: string
      permission_id:
        type: string
      permission_name:
        type: string
      permissions:
        $ref: '#/definitions/models.PermissionJson'
    type: object
...
  models.PermissionJson:
    properties:
      create:
        type: boolean
      delete:
        type: boolean
      edit:
        type: boolean
      view:
        type: boolean
    type: object
  models.Role:
    properties:
      created_at:
        description: วันที่และเวลาที่สร้างบทบาท
        type: string
      permissions:
        allOf:
        - $ref: '#/definitions/models.PermissionJson'
        description: สิทธิ์ที่อนุญาตในบทบาทนั้น (JSON)
      role_id:
        description: รหัสบทบาท (Primary Key)
        type: string
      role_name:
        description: ชื่อบทบาท (Unique Index)
        type: string
      updated_at:
        description: วันที่และเวลาที่อัปเดตบทบาท (Nullable)
        type: string
    type: object
  models.RolePermissionBody:
    properties:
      group_permission:
        items:
          $ref: '#/definitions/models.GroupRolePermission'
        type: array
      role_id:
        type: string
      role_name:
        type: string
      role_permission_id:
        type: string
    type: object
info:
  contact: {}
  description: เอกสารการใช้งาน API Land Registeation สำหรับ Developer
  title: Land Registeation API Documentation
  version: 0.0.9
paths:
...
  /role/:
    get:
      consumes:
      - application/json
      description: get GetallRoleEndpoint
      produces:
      - application/json
      responses:
        "200":
          description: role data
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all role
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: create  new role by request body
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.roleCreated'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            type: string
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: handler func  Create role endpoint
      tags:
      - Role
  /role/:role_id:
    get:
      consumes:
      - application/json
      description: put GetoneRoleEndpoint
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/models.Role'
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get one role
      tags:
      - Role
  /role/{role_id}:
    put:
      consumes:
      - application/json
      description: put EditRolePermissionEndpoint
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: Request body for editing an role permission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PermissionJson'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            type: string
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Edit permission role
      tags:
      - Role
  /role/{role_id}/permission_detail:
    put:
      consumes:
      - application/json
      description: put UpdateRolePermissionEndpoint
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: Request body for editing an role permission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RolePermissionBody'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            type: string
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update role permission detail
      tags:
      - RolePermission
  /role/role_permission/:
    post:
      consumes:
      - application/json
      description: post CreateRolePermissionEndpoint
      parameters:
      - description: Request body for created role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RolePermissionBody'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            type: string
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create role permission
      tags:
      - RolePermission
  /role_permission:
    get:
      consumes:
      - application/json
      description: get GetallRolePermissionEndPoint
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            items:
              $ref: '#/definitions/models.RolePermissionBody'
            type: array
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all role permission
      tags:
      - RolePermission
  /role_permission/{role_permiassion_id}:
    get:
      consumes:
      - application/json
      description: get GetRolePermissionEndPoint
      parameters:
      - description: Role ID
        in: path
        name: role_permiassion_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/models.RolePermissionBody'
        "401":
          description: Unauthorized error
          schema:
            type: string
        "404":
          description: Data not found error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get one role permission
      tags:
      - RolePermission
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"